requirements.txt
# Core AI/ML Libraries
langchain==0.1.0
langchain-community==0.0.20
langchain-openai==0.0.5
langgraph==0.0.20
openai>=1.12.0

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Vector Store and Retrieval
chromadb==0.4.22
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# Text Processing
nltk==3.8.1
spacy==3.7.2
whoosh==2.7.4

# Data Handling
pandas==2.2.0
numpy==1.26.3
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
rich==13.7.0
python-multipart==0.0.9
httpx==0.27.0

# Evaluation
scikit-learn==1.4.0
pytest==8.3.5
pytest-asyncio==0.25.3
rank-bm25==0.2.2

# Optional: For better PDF parsing
# Keep this aligned with pdfplumber's pinned dependency to avoid resolver conflicts.
pdfminer.six==20221105
